Ja so spät, dafür endboss szene läuft

+++ Assets/Scenes/Menu_Scenes/HighscoreWinnerScene.cs
@@ -10,7 +10,7 @@ public class HighscoreWinnerScene : MonoBehaviour
     {
         if (PlayerPrefs.HasKey("Highscore"))
         {
-            float time = PlayerPrefs.GetFloat("Highscore");
+            int time = (int) PlayerPrefs.GetFloat("Highscore");
             score.SetText(time + " seconds");
         }
     }
